provides an API to delete the subscription id both from the list of subscriptions ids of an app and from the routingTable
Dispatcher keeps a list of subscriptions for each app: {app1Id: [sub1Id, sub2Id}, app2Id: [sub1Id, sub2Id]}
when a subscription is deleted, the id of that subscription must also be deleted form the list on the dispatcher.
the subscrition Id must also be deleted from the routing table.